Extract out lib methods for materials
